Create 20190723-AALE-3rd-Session-Most-Salient-Points .md
This AALE 3rd Session Most Salient Points is a digest feedback. It is a recommended warmup reading, before tackling a more comprehensive feedback, to be found in the "Full Takeaway" version coming next. Same methodology to apply in principle for upcoming sessions as well as the 2 previous session feedbacks remaining to be migrated toward Github repositories here.